+++ applicationPorts/src/main/java/pl/zzpj/controller/TransactionUseCase.java
@@ -6,9 +6,9 @@ import java.math.BigDecimal;
 
 public interface TransactionUseCase {
 
-    void withdraw(Account account, double amount);
+    void withdraw(Account account, BigDecimal amount);
 
-    void deposit(Account account, double amount);
+    void deposit(Account account, BigDecimal amount);
 
-    void transfer(Account from, Account to, double amount, BigDecimal rate) throws Exception;
+    void transfer(Account from, Account to, BigDecimal amount, BigDecimal rate) throws Exception;
 }
